Module config
Now modules can have a config file when configured as `modules/name/module.ctp`
and `modules/name/config.ctp`. Previous modules still work without the config and
will be given the old text area for json config.

Added code for ajaxing the config in the backend as modules are changed.

Added method for generating the config forms

removed author details from the modules as they are only ever part of a plugin which
has the author details already.